+++ src/setup-docker/templates/fabric-docker.sh
@@ -12,7 +12,7 @@ source "$FABLO_NETWORK_ROOT/fabric-docker/snapshot-scripts.sh"
 source "$FABLO_NETWORK_ROOT/fabric-docker/commands-generated.sh"
 source "$FABLO_NETWORK_ROOT/fabric-docker/.env"
 
-function networkUp() {
+networkUp() {
   generateArtifacts
   startNetwork
   generateChannelsArtifacts
@@ -26,7 +26,7 @@ if [ "$1" = "up" ]; then
   networkUp
 elif [ "$1" = "down" ]; then
   networkDown
-elif [ "$1" = "reboot" ]; then
+elif [ "$1" = "reset" ]; then
   networkDown
   networkUp
 elif [ "$1" = "start" ]; then
@@ -49,7 +49,7 @@ elif [ "$1" = "--help" ]; then
   printHelp
 else
   echo "No command specified"
-  echo "Basic commands are: up, down, start, stop, reboot"
+  echo "Basic commands are: up, down, start, stop, reset"
   echo "To list channel query helper commands type: 'fablo channel --help'"
   echo "Also check: 'chaincode install'"
   echo "Use 'help' or '--help' for more information"