+++ src/setup-docker/templates/fabric-docker/scripts/base-functions.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-function certsGenerate() {
+certsGenerate() {
   local CONTAINER_NAME=certsGenerate
 
   local CONFIG_PATH=$1
@@ -18,7 +18,7 @@ function certsGenerate() {
 
   if [ -d "$FULL_CERT_PATH" ]; then
     echo "Can't generate certs, directory already exists : $FULL_CERT_PATH"
-    echo "Try using 'reboot' or 'down' to remove whole network or 'start' to reuse it"
+    echo "Try using 'reset' or 'down' to remove whole network or 'start' to reuse it"
     exit 1
   fi
 
@@ -38,7 +38,7 @@ function certsGenerate() {
   done
 }
 
-function genesisBlockCreate() {
+genesisBlockCreate() {
   local CONTAINER_NAME=genesisBlockCreate
 
   local CONFIG_PATH=$1
@@ -54,7 +54,7 @@ function genesisBlockCreate() {
 
   if [ -f "$OUTPUT_PATH/$GENESIS_FILE_NAME" ]; then
     echo "Cant't generate genesis block, file already exists: $OUTPUT_PATH/$GENESIS_FILE_NAME"
-    echo "Try using 'reboot' or 'down' to remove whole network or 'start' to reuse it"
+    echo "Try using 'reset' or 'down' to remove whole network or 'start' to reuse it"
     exit 1
   fi
 
@@ -70,7 +70,7 @@ function genesisBlockCreate() {
   removeContainer $CONTAINER_NAME
 }
 
-function createChannelTx() {
+createChannelTx() {
   local CONTAINER_NAME=createChannelTx
 
   local CHANNEL_NAME=$1
@@ -87,7 +87,7 @@ function createChannelTx() {
 
   if [ -f "$CHANNEL_TX_PATH" ]; then
     echo "Can't create channel configuration, it already exists : $CHANNEL_TX_PATH"
-    echo "Try using 'reboot' or 'down' to remove whole network or 'start' to reuse it"
+    echo "Try using 'reset' or 'down' to remove whole network or 'start' to reuse it"
     exit 1
   fi
 
@@ -102,7 +102,7 @@ function createChannelTx() {
   removeContainer $CONTAINER_NAME
 }
 
-function createNewChannelUpdateTx() {
+createNewChannelUpdateTx() {
   local CONTAINER_NAME=createAnchorPeerUpdateTx
 
   local CHANNEL_NAME=$1
@@ -122,7 +122,7 @@ function createNewChannelUpdateTx() {
 
   if [ -f "$ANCHOR_PEER_UPDATE_PATH" ]; then
     echo "Cant't create anchor peer update, it already exists : $ANCHOR_PEER_UPDATE_PATH"
-    echo "Try using 'reboot' or 'down' to remove whole network or 'start' to reuse it"
+    echo "Try using 'reset' or 'down' to remove whole network or 'start' to reuse it"
     exit 1
   fi
 
@@ -142,7 +142,7 @@ function createNewChannelUpdateTx() {
   removeContainer $CONTAINER_NAME
 }
 
-function notifyOrgAboutNewChannel() {
+notifyOrgAboutNewChannel() {
   local CHANNEL_NAME=$1
   local MSP_NAME=$2
   local CLI_NAME=$3
@@ -169,7 +169,7 @@ function notifyOrgAboutNewChannel() {
   fi
 }
 
-function notifyOrgAboutNewChannelTls() {
+notifyOrgAboutNewChannelTls() {
   local CHANNEL_NAME=$1
   local MSP_NAME=$2
   local CLI_NAME=$3
@@ -198,7 +198,7 @@ function notifyOrgAboutNewChannelTls() {
   fi
 }
 
-function deleteNewChannelUpdateTx() {
+deleteNewChannelUpdateTx() {
   local CHANNEL_NAME=$1
   local MSP_NAME=$2
   local CLI_NAME=$3
@@ -214,7 +214,7 @@ function deleteNewChannelUpdateTx() {
   fi
 }
 
-function printHeadline() {
+printHeadline() {
   bold=$'\e[1m'
   end=$'\e[0m'
 
@@ -223,7 +223,7 @@ function printHeadline() {
   printf "${bold}============ %b %s %b ==============${end}\n" "\\$EMOJI" "$TEXT" "\\$EMOJI"
 }
 
-function printItalics() {
+printItalics() {
   italics=$'\e[3m'
   end=$'\e[0m'
 
@@ -232,26 +232,26 @@ function printItalics() {
   printf "${italics}==== %b %s %b ====${end}\n" "\\$EMOJI" "$TEXT" "\\$EMOJI"
 }
 
-function inputLog() {
+inputLog() {
   end=$'\e[0m'
   darkGray=$'\e[90m'
 
   echo "${darkGray}   $1 ${end}"
 }
 
-function inputLogShort() {
+inputLogShort() {
   end=$'\e[0m'
   darkGray=$'\e[90m'
 
   echo "${darkGray}   $1 ${end}"
 }
 
-function certsRemove() {
+certsRemove() {
   local CERTS_DIR_PATH=$1
   rm -rf "$CERTS_DIR_PATH"
 }
 
-function removeContainer() {
+removeContainer() {
   CONTAINER_NAME=$1
   docker rm -f "$CONTAINER_NAME"
 }