+++ src/setup-docker/templates/fabric-docker/scripts/base-help.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-function printHelp() {
+printHelp() {
   echo "Fablo is powered by SoftwareMill"
 
   echo ""
@@ -22,7 +22,7 @@ function printHelp() {
   echo "./fabric-docker.sh stop"
   echo -e "\t Stops already running network."
   echo ""
-  echo "./fabric-docker.sh reboot"
+  echo "./fabric-docker.sh reset"
   echo -e "\t Fresh start - it destroys whole network, certs, configs and then reruns everything."
   echo ""
   echo "./fabric-docker.sh channel --help"